# src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Crypto Forge API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  jwt:
    type: apiKey
    name: "Authorization"
    in: header
paths:
  /organization/invite/{organizationId}/accept:
    x-swagger-router-controller: organizationInvite
    get:
      description: Request user accepts an invitation for an organization.
      operationId: acceptInvite
      security:
        - jwt: []
      x-security-scope:
        - admin
        - user
      parameters:
        - name: organizationId
          in: path
          description: id of organization
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: Error
          schema:
            $ref:
              '#/definitions/ErrorResponse'
  /organization/invite/{organizationId}/destroy:
    x-swagger-router-controller: organizationInvite
    delete:
      description: Request user deletes an invitation from an organization.
      operationId: destroyInvite
      security:
        - jwt: []
      x-security-scope:
        - admin
        - user
      parameters:
        - name: organizationId
          in: path
          description: id of organization
          required: true
          type: string
      responses:
        "204":
          description: Success
          schema:
            $ref:
              '#/definitions/SuccessResponse'
        default:
          description: Error
          schema:
            $ref:
              '#/definitions/ErrorResponse'
  /organization/invite:
    x-swagger-router-controller: organizationInvite
    get:
      description: Gets a list of all invites
      operationId: indexInvites
      security:
        - jwt: []
      x-security-scope:
        - admin
        - user
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              type: object
              properties:
                _id:
                  type: string
                name:
                  type: string
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    
    post:
      description: Invites a user to an organization
      operationId: createInvite
      security:
        - jwt: []
      x-security-scope:
        - admin
        - user
      parameters:
        - in: body
          name: body
          description: email and organization id
          required: true
          schema:
            $ref: "#/definitions/OrganizationInvite"
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
    OrganizationInvite:
      title: An invite to an organization
      type: object
      required:
        - email
        - organization
      properties:
        email:
          type: string
        organization:
          type: string
...
    ErrorResponse:
      required:
        - message
      properties:
        message:
          type: string
    SuccessResponse:
      required:
        - message
      properties:
        message:
          type: string